feat(tmrx-1479): tidy up component
