+++ packages/ts-newskit/src/slices/list-view-slice/artcile-stack.tsx
@@ -3,24 +3,18 @@ import {
   LeadArticle,
   LeadArticleProps
 } from '../../components/slices/lead-article';
-import { ClickHandlerType } from '../types';
 import { Divider, Block, Visible, Hidden, Stack } from 'newskit';
 import { StyledAdBlock, AdBlockWrapperMedium } from './styles';
-
-export interface ListViewSliceProps {
-  leadArticle: LeadArticleProps[];
-  clickHandler: ClickHandlerType;
-  articleWithAdSlot: LeadArticleProps;
-}
+import { ListViewSliceProps } from '.';
 
 export const ArticleStack = ({
-  leadArticle,
+  leadArticles,
   articleWithAdSlot,
   clickHandler
 }: ListViewSliceProps) => {
   return (
     <>
-      {leadArticle.map((item: LeadArticleProps, index: number) => {
+      {leadArticles.map((item: LeadArticleProps, index: number) => {
         return (
           <>
             <LeadArticle
@@ -39,8 +33,9 @@ export const ArticleStack = ({
               clickHandler={clickHandler}
             />
             <Visible md>
-              {leadArticle.length > 1 &&
-                index < leadArticle.length - 1 &&
+              {leadArticles.length > 1 &&
+                index < leadArticles.length - 1 &&
+                articleWithAdSlot &&
                 articleWithAdSlot.id !== item.id && (
                   <Block marginBlock="space040">
                     <Divider
@@ -52,8 +47,8 @@ export const ArticleStack = ({
                 )}
             </Visible>
             <Hidden md>
-              {leadArticle.length > 1 &&
-                index < leadArticle.length - 1 && (
+              {leadArticles.length > 1 &&
+                index < leadArticles.length - 1 && (
                   <Block marginBlock="space040">
                     <Divider
                       overrides={{
@@ -64,25 +59,26 @@ export const ArticleStack = ({
                 )}
             </Hidden>
             <AdBlockWrapperMedium>
-              {articleWithAdSlot.id === item.id && (
-                <Visible md>
-                  <Divider
-                    overrides={{
-                      marginBlockEnd: 'space030',
-                      marginBlockStart: 'space060',
-                      stylePreset: 'lightDivider'
-                    }}
-                  />
-                  <Stack
-                    flow="horizontal-center"
-                    stackDistribution="center"
-                    marginBlockEnd="space030"
-                  >
-                    ADVERTISEMENT
-                  </Stack>
-                  <StyledAdBlock />
-                </Visible>
-              )}
+              {articleWithAdSlot &&
+                articleWithAdSlot.id === item.id && (
+                  <Visible md>
+                    <Divider
+                      overrides={{
+                        marginBlockEnd: 'space030',
+                        marginBlockStart: 'space060',
+                        stylePreset: 'lightDivider'
+                      }}
+                    />
+                    <Stack
+                      flow="horizontal-center"
+                      stackDistribution="center"
+                      marginBlockEnd="space030"
+                    >
+                      ADVERTISEMENT
+                    </Stack>
+                    <StyledAdBlock />
+                  </Visible>
+                )}
             </AdBlockWrapperMedium>
           </>
         );
