+++ packages/ts-newskit/src/slices/list-view-slice/styles.ts
@@ -3,7 +3,7 @@ import {
   styled,
   getMediaQueryFromTheme,
   TextBlock,
-  PaginationButton,
+  PaginationButton
 } from 'newskit';
 
 export const StyledBlock = styled(Block)`
